Clarify VKS upload API
The response for the upload API appears to make it clear that it is
intended to accept only a single OpenPGP certificate, not multiple
keys.

I don't know for sure whether this is the intent of the codebase; if
it is not, then the API is broken, and the response for the upload API
needs to be able to accomodate multiple primary key fingerprints.