instruct --locate-keys to use the pre-defined keyserver
By default, --locate-keys uses the key retrieval mechanisms
"local,wkd".  If you want it to use the keyserver, you need to tell it
to do so explicitly.

An alternate approach would be to add the following to gpg.conf:

    auto-key-locate local,wkd,hkps://keys.openpgp.org

but if you did that, you'd be getting into policy choices about
metadata leakage.